Merged: Merge changes r5487 - r5508 from the trunk.
git-svn-id: svn://svn.webtrees.net/branches/phpGedView-ca@5510 4201c862-f99b-1943-94d0-25fd05010cbc

+++ inverselink.php
@@ -139,14 +139,20 @@ if ($action == "choose" && $paramok) {
 	print "<tr><td class=\"descriptionbox\">";
 	
 	if ($linkto == "person") {
-		// print $pgv_lang["enter_pid"]."</td>";
-		echo "Add more links:";
+		if (file_exists('modules/GEDFact_assistant/MEDIA/media_1_ctrl.php')) {
+			echo "Add more links:";
+		} else{
+			print $pgv_lang["enter_pid"]."</td>";
+		}
 		print "<td class=\"optionbox wrap\">";
-		echo "A)&nbsp;&nbsp;";
+		
+		// GEDFact assistant Add Media Links =======================
+		if (file_exists('modules/GEDFact_assistant/MEDIA/media_1_ctrl.php')) {
+			echo "First, enter a Base individual ID<br />";
+		}
 		if ($linktoid=="") {
 			print "<input class=\"pedigree_form\" type=\"text\" name=\"linktoid\" id=\"linktopid\" size=\"3\" value=\"$linktoid\" />";
 			print_findindi_link("linktopid","");
-		echo "&nbsp;&nbsp;".$pgv_lang["enter_pid"];
 		} else {
 			$record=Person::getInstance($linktoid);
 			echo '<b>', PrintReady($record->getFullName()), '</b>&nbsp;&nbsp;&nbsp;';
@@ -154,16 +160,23 @@ if ($action == "choose" && $paramok) {
 			print "(".$linktoid.")";
 			if ($TEXT_DIRECTION=="rtl") print getRLM();
 		}
-		echo "<br /><br />";
+		
+		// GEDFact assistant Add Media Links =======================
+		if (file_exists('modules/GEDFact_assistant/MEDIA/media_1_ctrl.php')) {
+			echo "<br /><br />";
+			echo "Then, click Add to add more Individual Links";
+			echo "<br />";
+		}
+		
 		// GEDFact assistant Add Media Links =======================
 		if (file_exists('modules/GEDFact_assistant/MEDIA/media_1_ctrl.php')) {
 			include ('modules/GEDFact_assistant/MEDIA/media_query_2a.php');
 			echo "</td></tr>";
 			echo "<tr><td class=\"topbottombar\" colspan=\"2\">";
-			echo "<input type=\"button\" value=\"".$pgv_lang["set_link"]."s\" onclick=\"javascript:alert('Clicking \'Set Links\' will eventually parse and save the Current and Added Links');\" />";
+			echo "<input type=\"submit\" value=\"".$pgv_lang["set_link"]."s\" onclick=\"javascript:alert('Clicking \'Set Links\' will eventually parse and save the Current and Added Links, for now only the \'Base\' id will be saved');\" />";
 			echo "</td></tr>";
-		}else{
-		print "<tr><td class=\"topbottombar\" colspan=\"2\"><input type=\"submit\" value=\"".$pgv_lang["set_link"]."\" /></td></tr>";
+		} else {
+			print "<tr><td class=\"topbottombar\" colspan=\"2\"><input type=\"submit\" value=\"".$pgv_lang["set_link"]."\" /></td></tr>";
 		}
 
 	}
